Authorization complete

+++ app/Http/Controllers/Backend/RoleController.php
@@ -7,6 +7,7 @@ use App\Models\Module;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
@@ -17,6 +18,7 @@ class RoleController extends Controller
      */
     public function index()
     {
+        Gate::authorize('app.role.index');
         $roles = Role::all();
         return view('backend.roles.index', compact('roles'));
     }
@@ -28,6 +30,7 @@ class RoleController extends Controller
      */
     public function create()
     {
+        Gate::authorize('app.role.create');
         $modules = Module::all();
         return view('backend.roles.form', compact('modules'));
     }
@@ -40,6 +43,7 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('app.role.create');
         $this->validate($request, [
             'name' => 'required|unique:roles',
             'permissions' =>'required|array',
@@ -74,6 +78,7 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
+        Gate::authorize('app.role.edit');
         $modules = Module::all();
         return view('backend.roles.form', compact('modules', 'role'));
     }
@@ -87,6 +92,7 @@ class RoleController extends Controller
      */
     public function update(Request $request, Role $role)
     {
+        Gate::authorize('app.role.edit');
         $role->update([
             'name' => $request->name,
             'slug' =>Str::slug($request->name)
@@ -105,6 +111,7 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        Gate::authorize('app.role.delete');
         if($role->deletable){
             $role->delete();
             notify()->success("Role Deleted","Success");
